requirements.txt
# FastAPI Backend
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Database
supabase>=2.3.0

# AI & LLM
openai>=1.30.0
sentence-transformers>=2.0.0
torch>=2.0.0

# Utilities
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv==1.0.0
tenacity==8.2.3

# Testing
pytest==7.4.3
httpx==0.27.2

# HTTP Requests
requests==2.31.0